LayoutTests/imported/w3c/web-platform-tests/fs/FileSystemSyncAccessHandle-truncate.https.tentative.worker-expected.txt


FAIL test SyncAccessHandle.truncate with different sizes assert_equals: expected (number) 4 but got (object) object "[object Promise]"
FAIL test SyncAccessHandle.truncate after SyncAccessHandle.write promise_test: Unhandled rejection with value: object "InvalidStateError: The object is in an invalid state."
